Use a constant for delimiter and update db filter.
