+++ ui/pages/Project/components/SavedVariants.jsx
@@ -15,6 +15,7 @@ import {
   VARIANT_PER_PAGE_FIELD,
   EXCLUDED_TAG_NAME,
   REVIEW_TAG_NAME,
+  TAG_URL_DELIMITER,
 } from 'shared/utils/constants'
 import UpdateButton from 'shared/components/buttons/UpdateButton'
 import { LargeMultiselect, Dropdown } from 'shared/components/form/Inputs'
@@ -150,7 +151,7 @@ class BaseProjectSavedVariants extends React.PureComponent {
     return getSavedVariantsLinkPath({
       projectGuid: project.projectGuid,
       analysisGroupGuid: match.params.analysisGroupGuid,
-      tag: !isCategory && lastNewTag !== ALL_FILTER && (updatedTag || []).join(';'),
+      tag: !isCategory && lastNewTag !== ALL_FILTER && (updatedTag || []).join(TAG_URL_DELIMITER),
       familyGuid: match.params.familyGuid,
     })
   }
